06/6.5/requirements.txt
# requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
PyJWT>=2.8.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
slowapi>=0.1.8